app/ml/face_swapper/requirements.txt
dlib==19.24.1
firebase_admin==6.1.0
Flask==2.3.2
numpy==1.23.5
opencv_python==4.7.0.72
python-dotenv==1.0.0
